Buncha SEP-related module/file renames
