+++ src/soroban/index.ts
@@ -7,7 +7,6 @@ export * from './soroban_rpc';
 // soroban-client classes to expose
 export { Server, Durability } from './server';
 export { default as AxiosClient } from './axios';
-export { ContractSpec } from './contract_spec';
 export { parseRawSimulation, parseRawEvents } from './parsers';
 export * from './transaction';
